fixing case where 0 octets throw exception
This was mentioned in the comments of this article:

https://www.mikemackintosh.com/5-tips-for-working-with-ipv6-in-php/

Specifically:

>In those functions, the 'A' causes '0' octets to become an ASCII space
(32 or \x20), rather than the NUL that an 'a' produces.

I setup a couple of tests that exhibited this behavior, which would
throw this exception when the IP had a “0” for an octet:

>Exception: Please provide a 4 or 16 byte string
>
>/Users/jasonklehr/Sites/dTR-IP/src/dTR/Networking/dTRIP.php:231

Changing the “A” in the pack/unpack calls to “a” seems to fix this
behavior.

I’ve added a few other ranges to the test data provider for good
measure.